made code cleaner, new files for functions that were previously in the main mjs files for pages, organised files better for clarity.
