+++ index.html
@@ -80,9 +80,8 @@
                     role="button"
                     data-bs-toggle="dropdown"
                     aria-expanded="false"
-                    id="userCredits"
                   >
-                    Credits: <span id="userCredits">0</span>
+                    <span id="userDisplay"></span>
                   </a>
                   <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
                     <li>
