+++ src/js/api/listings/handle-create-listing.mjs
@@ -1,0 +1,46 @@
+import { makeApiRequest } from '../api-service.mjs';
+import { displayError } from '../../utilities/messages/error-handler.mjs';
+import { displaySuccess } from '../../utilities/messages/success.mjs';
+
+export async function handleCreateListingForm(event) {
+  event.preventDefault();
+  const title = document.getElementById('title').value.trim();
+  const description = document.getElementById('description').value.trim();
+  const tags = document
+    .getElementById('tags')
+    .value.trim()
+    .split(',')
+    .map((tag) => tag.trim());
+  const mediaInputs = document.querySelectorAll('#mediaContainer input');
+  const media = Array.from(mediaInputs)
+    .map((input) => input.value.trim())
+    .filter((url) => url);
+
+  const endsAt = document.getElementById('endsAt').value;
+
+  try {
+    const response = await makeApiRequest('listings', 'POST', {
+      title,
+      description,
+      tags,
+      media,
+      endsAt,
+    });
+
+    const listingUrl = `/src/html/listing-specific/index.html?id=${encodeURIComponent(
+      response.id
+    )}`;
+    displaySuccess('Listing created successfully!', listingUrl, () => {
+      const link = document.querySelector('#successMessage .alert-link');
+      if (link) {
+        link.addEventListener('click', () => {
+          console.log('Link clicked');
+        });
+      }
+    });
+    document.getElementById('createListingForm').reset();
+    console.log('Listing created:', response);
+  } catch (error) {
+    displayError();
+  }
+}
