+++ src/js/pages/create-listing.mjs
@@ -1,15 +1,14 @@
 import '../userstate-display/logged-in-visible.mjs';
 import { setUpLogoutLink } from '../api/authenticate/logout.mjs';
-import { setupCreateListingForm } from '../api/listings/setup-create-listing-form.mjs';
-import { makeApiRequest } from '../api/api-service.mjs';
-import { displayError } from '../utilities/messages/error-handler.mjs';
+import { setupCreateListingForm } from '../setup/setup-create-listing-form.mjs';
 import { updateUserCredits } from '../utilities/update-credit.mjs';
-import { displaySuccess } from '../utilities/messages/success.mjs';
+import { handleCreateListingForm } from '../api/listings/handle-create-listing.mjs';
+import { setUpSearchForm } from '../setup/set-up-search.mjs';
 
 document.addEventListener('DOMContentLoaded', () => {
   setUpLogoutLink();
-
   updateUserCredits();
+  setUpSearchForm();
 
   const createListingContainer = document.getElementById(
     'createListingContainer'
@@ -18,47 +17,6 @@ document.addEventListener('DOMContentLoaded', () => {
     setupCreateListingForm();
     document
       .getElementById('createListingForm')
-      .addEventListener('submit', async (event) => {
-        event.preventDefault();
-        const title = document.getElementById('title').value.trim();
-        const description = document.getElementById('description').value.trim();
-        const tags = document
-          .getElementById('tags')
-          .value.trim()
-          .split(',')
-          .map((tag) => tag.trim());
-        const mediaInputs = document.querySelectorAll('#mediaContainer input');
-        const media = Array.from(mediaInputs)
-          .map((input) => input.value.trim())
-          .filter((url) => url);
-
-        const endsAt = document.getElementById('endsAt').value;
-
-        try {
-          const response = await makeApiRequest('listings', 'POST', {
-            title,
-            description,
-            tags,
-            media,
-            endsAt,
-          });
-
-          const listingUrl = `/src/html/listing-specific/index.html?id=${encodeURIComponent(
-            response.id
-          )}`;
-          displaySuccess('Listing created successfully!', listingUrl, () => {
-            const link = document.querySelector('#successMessage .alert-link');
-            if (link) {
-              link.addEventListener('click', () => {
-                console.log('Link clicked');
-              });
-            }
-          });
-          document.getElementById('createListingForm').reset();
-          console.log('Listing created:', response);
-        } catch (error) {
-          displayError();
-        }
-      });
+      .addEventListener('submit', handleCreateListingForm);
   }
 });
