+++ src/js/utilities/update-credit.mjs
@@ -1,21 +1,20 @@
 import { getUserInfo } from '../storage/storage.mjs';
-import { getUserProfile } from '../api/profile/profile-fetch.mjs';
+import { getUserProfile } from '../api/profile/user-profile-API.mjs';
 
 export async function updateUserCredits() {
   const userInfo = getUserInfo();
   if (userInfo) {
     const userProfile = await getUserProfile(userInfo.name);
-    console.log('Updated profile', userProfile);
-    const creditsElement = document.getElementById('userCredits');
-    if (creditsElement) {
-      creditsElement.innerHTML = `<span class="material-symbols-outlined align-middle fs-5">
+    const userDisplayElement = document.getElementById('userDisplay');
+    if (userDisplayElement) {
+      userDisplayElement.innerHTML = `${userInfo.name} <span class="material-symbols-outlined align-middle ps-2 pb-1">
       universal_currency_alt
       </span> ${userProfile.credits}`;
     }
   } else {
-    const creditsElement = document.getElementById('userCredits');
-    if (creditsElement) {
-      creditsElement.textContent = '';
+    const userDisplayElement = document.getElementById('userDisplay');
+    if (userDisplayElement) {
+      userDisplayElement.textContent = 'Credits: 0';
     }
   }
 }
